# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API V1",
        "version": "0.1"
    },
    "paths": {
        "api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new User",
                "operationId": "be6599bdfc41183cf0d19b2519f60ea9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": 0
                                        },
                                        "message": {
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": {
                                                "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login an existing User",
                "operationId": "4ce4077807072036e82d5da69026cf3a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": 0
                                        },
                                        "message": {
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": {
                                                "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/v1/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Retrive a listing of the Articles",
                "operationId": "268a03dbb146ec4695d2f80c4228e66e",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer epl5d5olRkge9DK60acfBrrFIHufNeVIXngSWJ7ReCNkr11I6WL"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": 200
                                        },
                                        "message": {
                                            "example": "OK"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Article"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Store new Article",
                "operationId": "bb4d19dc1d2e9966e5af449dfff3c44a",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer epl5d5olRkge9DK60acfBrrFIHufNeVIXngSWJ7ReCNkr11I6WL"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreArticleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": 201
                                        },
                                        "message": {
                                            "example": "OK"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Article"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/v1/articles/{uuid}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get an existing Article",
                "operationId": "19ce0ac1439cf3881962b4ffdb550044",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Article uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "389ffffe-b89c-47b6-bc63-cf5fd2a88218"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer epl5d5olRkge9DK60acfBrrFIHufNeVIXngSWJ7ReCNkr11I6WL"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": 200
                                        },
                                        "message": {
                                            "example": "OK"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Article"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Delete existing Article",
                "operationId": "8b0f4c6db5f6afe171e96c2b943fb46b",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Article uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "389ffffe-b89c-47b6-bc63-cf5fd2a88218"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer epl5d5olRkge9DK60acfBrrFIHufNeVIXngSWJ7ReCNkr11I6WL"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "example": 204
                                        },
                                        "message": {
                                            "example": "OK"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Article"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "title": "LoginRequest",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "example@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "Password1!"
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "title": "RegisterRequest",
                "properties": {
                    "username": {
                        "type": "string",
                        "example": "my-username"
                    },
                    "email": {
                        "type": "string",
                        "example": "example@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "Password1!"
                    }
                },
                "type": "object"
            },
            "StoreArticleRequest": {
                "title": "StoreArticleRequest",
                "properties": {
                    "content": {
                        "type": "string",
                        "example": "text based article"
                    }
                },
                "type": "object"
            },
            "Article": {
                "title": "Article",
                "properties": {
                    "uuid": {
                        "type": "string",
                        "example": "389ffffe-b89c-47b6-bc63-cf5fd2a88218"
                    },
                    "content": {
                        "type": "string",
                        "example": "example@example.com"
                    },
                    "type": {
                        "type": "string",
                        "example": "Text"
                    },
                    "result": {
                        "type": "string",
                        "example": "example@example.com"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "example@example.com"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "properties": {
                    "uuid": {
                        "type": "string",
                        "example": "389ffffe-b89c-47b6-bc63-cf5fd2a88218"
                    },
                    "username": {
                        "type": "string",
                        "example": "my-username"
                    },
                    "email": {
                        "type": "string",
                        "example": "example@example.com"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Articles",
            "description": "Articles"
        }
    ]
}